// src/app/models/interfaces/userGoogle-response.interface.ts
export interface UserGoogleResponse {
  at_hash: string
  aud: string
  azp: string
  email: string
  email_verified: boolean
  exp: number
  family_name: string
  given_name: string
  iat: number
  iss: string
  jti: string
  name: string
  nbf: number
  nonce: string
  picture: string
  sub: string
}